requirements.txt
django
gunicorn
Pillow
django-object-actions
django-mptt
requests
djlint
django-environ
django-solo